Create jwtauthenticationfile
Go to the Box Developer Console and create a new JWT application.
After creating the application, generate a private key and public key pair for the application.
Now, you can use the following code to generate a new JWT:

Now, you can use the generated JWT to authenticate with the Box API and renew the certificate:

Make sure to replace the placeholder values with your actual Box application details. If you want to renew additional certificates, you will need to generate a new JWT for each certificate and make an API call for each one.




How can I renew multiple certificates at once


To renew multiple certificates at once, you can use the same JWT to authenticate with the Box API and then use the Box API to list all certificates for your enterprise.

After obtaining the list of certificates, you can then loop through the list and renew each certificate using the Box API.

In this example, make sure to replace 'YOUR_BOX_USER_ID' with the actual Box user ID that you want to use for the API calls.

Note that this code assumes that all the certificates need to be renewed to the same date ('2023-01-01T00:00:00Z' in this case). If you want to renew the certificates to different dates, you will need to modify the 'expires_at' field in the 'certificate_renewal_payload' accordingly.